Crate s3lsio

Structs

Client
Client structure holds a reference to the S3Client which also implements two traits: AwsCredentialsProvider and DispatchSignedRequest Since S3Client struct is takes those two traits as parameters then ALL functions called that require passing in S3Client or Client must specify the trait signature as follows: Example: fn whatever_function(client: &mut Client) Note: Could also specify 'where' P:... D:... instead.

Error
Allows you to control Error output.

Output
Allows you to control non-Error output.

Enums

OutputFormat
Allows you to set the output type for stderr and stdout.