Struct s3lsio::BenchHostInstanceSummary

pub struct BenchHostInstanceSummary {
    pub host_instance: String,
    pub ip_address: String,
    pub start_time: String,
    pub end_time: String,
    pub total_requests: u64,
    pub total_success: u64,
    pub total_errors: u64,
    pub total_duration: f64,
    pub total_payload: u64,
    pub total_throughput: f64,
    pub total_threads: u64,
    pub host_duration: f64,
    pub operations: Vec<BenchThreadSummary>,
}

A summary of all threads on a given node/instance

NOTE: The start and end times DO NOT reflect a true duration. They represent the overall time block to execute the operations and compute the results for the given node/instance. Total_duration is the number of seconds.nanoseconds of actual execution time.

NOTE: Simple types are used here to make serialization easy. start and and end time DO NOT reflect a true duration. The total_duration does that.

Fields
host_instance: String ip_address: String start_time: String end_time: String total_requests: u64 total_success: u64 total_errors: u64 total_duration: f64 total_payload: u64 total_throughput: f64 total_threads: u64 host_duration: f64 operations: Vec<BenchThreadSummary>
Methods

impl BenchHostInstanceSummary

fn new(operations: Vec<BenchThreadSummary>) -> BenchHostInstanceSummary