Struct s3lsio::Client

pub struct Client<'a, P: 'a, D: 'a> where P: AwsCredentialsProvider, D: DispatchSignedRequest {
    pub s3client: &'a mut S3Client<P, D>,
    pub config: &'a mut Config,
    pub error: Error,
    pub output: Output,
    pub is_quiet: bool,
    pub is_time: bool,
    pub is_bench: bool,
    pub is_admin: bool,
}

Client structure holds a reference to the S3Client which also implements two traits: AwsCredentialsProvider and DispatchSignedRequest Since S3Client struct is takes those two traits as parameters then ALL functions called that require passing in S3Client or Client must specify the trait signature as follows: Example: fn whatever_function(client: &mut Client) Note: Could also specify 'where' P:... D:... instead.

Fields
s3client: &'a mut S3Client<P, D> config: &'a mut Config error: Error output: Output is_quiet: bool is_time: bool is_bench: bool is_admin: bool