Struct s3lsio::Output

pub struct Output {
    pub format: OutputFormat,
    pub color: Color,
}

Allows you to control non-Error output.

Fields
format: OutputFormat

Defaults to OutputFormat::plain.

Available formats are json, plain, serialize or none (don't output anything). If plain is used then you can serialize structures with format! and then pass the output.

color: Color

Can be any term color. Defaults to term::color::GREEN.