Crate s3lsio

If you want a configuration file to store options so that you don't want to pass those in each time then create a subdirectory in your home directory: mkdir ~/.s3lsio Create a TOML file called config: vim ~/.s3lsio/config Add the following options (optional): [options] endpoint = "" proxy = "" signature = "V4"

NOTE: You can set signature to V2 or V4 depending on the product you are going after. By default AWS S3 uses V4 but products like Ceph (Hammer release) use V2. Ceph (Jewel release) uses V4. The default is V4.

Structs

Client
Client structure holds a reference to the S3Client which also implements two traits: AwsCredentialsProvider and DispatchSignedRequest Since S3Client struct is takes those two traits as parameters then ALL functions called that require passing in S3Client or Client must specify the trait signature as follows: Example: fn whatever_function(client: &mut Client) Note: Could also specify 'where' P:... D:... instead.

Error
Allows you to control Error output.

Output
Allows you to control non-Error output.

Enums

Commands
Commands

OutputFormat
Allows you to set the output type for stderr and stdout.